+++ packages/implementations/react/src/atomic/Input/src/components/Search/Search.tsx
@@ -33,7 +33,7 @@ const InputSearch = forwardRef<HTMLInputElement, _InputSearchProps>(
     return (
       <div
         data-testid={dataTestid ? `${dataTestid}-container` : ""}
-        className={input.classnames.container({ appearance, border })}
+        className={input.classnames.container({ appearance, border, size })}
       >
         <input
           {...rest}